Peg the specific grunt-contrib-FOO versions.
We need to avoid installing grunt-contrib-jasmine
because there's a race condition installing that along with
grunt-contrib-qunit. They both install phantomjs in an async
manner, downloading in the background.
If the second one starts while the first one is still downloading,
it sees the phantomjs file and assumes it can just install it,
even if it's only N% of the way through the download.