Do not show "pretty" annotations on .wym-editor-only elements
These elements are not intended as part of the content to be edited, so
there is no point in showing annotations on them.

In particular, the annotations were previously added on the image resize
handle, partly obscuring it from view and preventing the mouse pointer
from reaching the handle, making image resizing impossible using the
pretty theme.

This fixes #784.